add Applicative instances
The instances for Error, State and RWS require that the argument is a Monad.
Unfortunately Functor is not a superclass of Monad but is a superclass of
Applicative, so the Functor instances for these constructors are back to
depending on Monad.

darcs-hash:0a05653cb419f9ab2f2b15135e4ca4c2d9ba650c