Deprecate Control.Monad.Trans.List
Ignore-this: 803a54f83f75fb1282172857c0451223

The ListT in this module is not a monad in the general case, requiring a
tricky invariant (commutativity in the base monad) to use properly.

There are plenty of packages which provide a "ListT done right", and we
should encourage new users to depend on those instead.

darcs-hash:8b4a90d92c8802e90ef9b146a6ebdf616f2abce4